+++ files/initrd/opt/arc/include/update.sh
@@ -65,22 +65,26 @@ function upgradeLoader () {
       if [ -f "${TMP_PATH}/arc.img.zip" ]; then
         echo "Downloading Upgradefile successful!"
       else
-        echo "Error downloading Upgradefile!"
-        sleep 5
-        return 1
+        updateFailed
       fi
       unzip -oq "${TMP_PATH}/arc.img.zip" -d "${TMP_PATH}"
       rm -f "${TMP_PATH}/arc.img.zip" >/dev/null
       echo "Installing new Loader Image..."
       # Process complete update
       umount "${PART1_PATH}" "${PART2_PATH}" "${PART3_PATH}"
-      if [ -n "${ARCBRANCH}" ]; then
-        dd if="${TMP_PATH}/arc-${ARCBRANCH}.img" of=$(blkid | grep 'LABEL="ARC3"' | cut -d3 -f1) bs=1M conv=fsync
+      if [ "${ARCBRANCH}" != "stable" ]; then
+        if dd if="${TMP_PATH}/arc-${ARCBRANCH}.img" of=$(blkid | grep 'LABEL="ARC3"' | cut -d3 -f1) bs=1M conv=fsync; then
+          rm -f "${TMP_PATH}/arc-${ARCBRANCH}.img" >/dev/null
+        else
+          updateFailed
+        fi
       else
-        dd if="${TMP_PATH}/arc.img" of=$(blkid | grep 'LABEL="ARC3"' | cut -d3 -f1) bs=1M conv=fsync
+        if dd if="${TMP_PATH}/arc.img" of=$(blkid | grep 'LABEL="ARC3"' | cut -d3 -f1) bs=1M conv=fsync; then
+          rm -f "${TMP_PATH}/arc.img" >/dev/null
+        else
+          updateFailed
+        fi
       fi
-      # Ask for Boot
-      rm -f "${TMP_PATH}/arc.img" >/dev/null
       echo "Upgrade done! -> Rebooting..."
       sleep 2
     ) 2>&1 | dialog --backtitle "$(backtitle)" --title "Upgrade Loader" \
@@ -596,7 +600,7 @@ function updateFailed() {
     exit 1
   else
     echo "Update failed!"
-    exit 1
+    return 1
   fi
 }
 
@@ -611,6 +615,6 @@ function updateFaileddialog() {
   else
     dialog --backtitle "$(backtitle)" --title "Update Failed" \
       --msgbox "Update failed!" 0 0
-    exit 1
+    return 1
   fi
 }
